agentic_backtesting/requirements.txt
backtrader>=1.9.78
pandas>=1.5.0
numpy>=1.24.0
yfinance>=0.2.0
python-dotenv>=1.0.0
streamlit>=1.28.0
matplotlib>=3.5.0
plotly>=5.0.0
psycopg2-binary>=2.9.0
strands-agents
strands-agents-tools
